chore: use criterion for benchmarks
This replaces rust's built in Bencher with `criterion` crate and allows us to stay on rust stable instead of having to toggle to nightly to run benchmarks. `criterion` also gives us some nice output.

Here is the peek_n benchmark, for example:
```bash
parser_peek_n           time:   [18.470 us 18.552 us 18.643 us]
                        change: [-0.0240% +0.3922% +0.8940%] (p = 0.09 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  8 (8.00%) high mild
  1 (1.00%) high severe
```
