+++ crates/apollo-parser/Cargo.toml
@@ -15,3 +15,9 @@ rowan = "0.13.0-pre.7"
 pretty_assertions = "0.7.1"
 expect-test = "1.1"
 unindent = "0.1.7"
+criterion = "0.3.0"
+
+[[bench]]
+name = "benches"
+path = "benches/benches.rs"
+harness = false
